third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-console/debug-console-command.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script src="../../../inspector/console-test.js"></script>

<script>
function simpleTestFunction()
{
   return 0;
}
</script>

<script>
function simpleTestFunction1() { return 0; } function simpleTestFunction2() { return 0; }
</script>

<script>
function simpleTestFunction3() { Math.random(); debugger; }
</script>

<script>
var test = function() {
  var currentSourceFrame;
  SourcesTestRunner.setQuiet(true);
  SourcesTestRunner.runDebuggerTestSuite([
    function testSetSimpleBreakpoint(next) {
      setBreakpointAndRun(next, 'simpleTestFunction', 'simpleTestFunction();');
    },

    function testSetBreakpointOnFirstFunctionInLine(next) {
      setBreakpointAndRun(next, 'simpleTestFunction1', 'simpleTestFunction2(); simpleTestFunction1();');
    },

    function testSetBreakpointOnLastFunctionInLine(next) {
      setBreakpointAndRun(next, 'simpleTestFunction2', 'simpleTestFunction1(); simpleTestFunction2();');
    },

    function testRemoveBreakpoint(next) {
      ConsoleTestRunner.evaluateInConsole('debug(simpleTestFunction3); undebug(simpleTestFunction3);');
      ConsoleTestRunner.evaluateInConsole('setTimeout(simpleTestFunction3, 0)');
      SourcesTestRunner.waitUntilPaused(didPause1);

      function didPause1(callFrames, reason) {
        TestRunner.addResult('Script execution paused.');
        TestRunner.addResult(
            'Reason for pause: ' +
            (reason == SDK.DebuggerModel.BreakReason.DebugCommand ? 'debug command' : 'debugger statement') + '.');
        next();
      }
    }
  ]);

  function setBreakpointAndRun(next, functionName, runCmd) {
    ConsoleTestRunner.evaluateInConsole('debug(' + functionName + ')');

    TestRunner.addResult('Breakpoint added.');
    ConsoleTestRunner.evaluateInConsole('setTimeout(function() { ' + runCmd + ' }, 0)');
    TestRunner.addResult('Set timer for test function.');
    SourcesTestRunner.waitUntilPaused(didPause);

    function didPause(callFrames, reason) {
      TestRunner.addResult('Script execution paused.');
      SourcesTestRunner.captureStackTrace(callFrames);
      ConsoleTestRunner.evaluateInConsole('undebug(' + functionName + ')');
      TestRunner.addResult('Breakpoint removed.');
      TestRunner.assertEquals(reason, SDK.DebuggerModel.BreakReason.DebugCommand);
      SourcesTestRunner.resumeExecution(didResume);
    }

    function didResume() {
      TestRunner.addResult('Script execution resumed.');
      next();
    }
  }
};

</script>
</head>

<body onload="runTest()">
<p>
Tests debug(fn) console command.
</p>

</body>
</html>
