third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-console/debug-console-command-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests debug(fn) console command.


Running: testSetSimpleBreakpoint
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction (debug-console-command.js:13)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testSetBreakpointOnFirstFunctionInLine
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction1 (debug-console-command.js:17)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testSetBreakpointOnLastFunctionInLine
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction2 (debug-console-command.js:17)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testRemoveBreakpoint
Script execution paused.
Reason for pause: debugger statement.
